Fix Cannot read property 'killed' of undefined
Fixes #679

+++ src/app/services/streaming/provider/validate.js
@@ -81,7 +81,7 @@ export default async function( execObj, providerConfData ) {
 		setTimeout( () => reject( new Error( "Timeout" ) ), validationTimeout );
 	})
 		.finally( () => {
-			if ( !child.killed ) {
+			if ( child && !child.killed ) {
 				child.kill( "SIGKILL" );
 				child = null;
 			}
